restructuring of files and some other improvements
