+++ .gitignore
@@ -1,0 +1,2 @@
+slaj
+slaj.exe
